runtime: return apis to separate file
